details.top div removed from details plate
