+++ components/Admin/Details/DetailsPlate/DetailsPlate.jsx
@@ -4,8 +4,9 @@ import BackgroundCanvas from "../../Shared/BackgroundCanvas/BackgroundCanvas";
 import TextParagraph from "../../Shared/Typography/TextParagraph";
 import TextBold from "../../Shared/Typography/TextBold";
 import { dateFormat } from "@/utils/serialize";
-import StatusCard from "../StatusCard/StatusCard";
+
 import { Pagination } from "@/utils/serialize";
+import StatusCard from "../../Details/StatusCard";
 
 export default function DetailsPlate({
   id,
@@ -51,87 +52,78 @@ export default function DetailsPlate({
   );
 
   return (
-    <div className={styles.top}>
-      <BackgroundCanvas className={styles.detailsMain}>
-        <div className={styles.top}>
-          <div className={styles.id}>
-            <div className={`${styles.idInner} ${styles.idSpacing}`}>
-              <TextBold variant="hash">#</TextBold>
-              <TextBold variant={styles.idLargeScreen}>id</TextBold>{" "}
-            </div>
-
-            <TextParagraph className={styles.header}>{id}</TextParagraph>
+    <BackgroundCanvas className={styles.detailsMain}>
+      <div className={styles.top}>
+        <div className={styles.id}>
+          <div className={`${styles.idInner} ${styles.idSpacing}`}>
+            <TextBold variant="hash">#</TextBold>
+            <TextBold variant={styles.idLargeScreen}>id</TextBold>{" "}
           </div>
 
-          <div className={`${styles.sub} ${styles.right}`}>
-            <TextParagraph
-              className={`${styles.defaultSpacing} ${styles.rightSide}`}
-            >
-              Submited
-            </TextParagraph>
-            <TextBold className={styles.largeSpacing}>
-              {dateFormat(createdAt)}
-            </TextBold>
-          </div>
-        </div>
-        <div className={styles.bottom}>
-          <div>
-            <TextParagraph className={styles.defaultSpacing}>By</TextParagraph>
-            <TextBold className={styles.largeSpacing}>{userName}</TextBold>
-          </div>
-          <div className={`${styles.sub} ${styles.right}`}>
-            <TextParagraph
-              className={`${styles.defaultSpacing} ${styles.rightSide}`}
-            >
-              Location
-            </TextParagraph>
-            <TextBold className={styles.largeSpacing}>{location}</TextBold>
-          </div>
+          <TextParagraph className={styles.header}>{id}</TextParagraph>
         </div>
 
-        <TextParagraph className={styles.smallerSpacing}>
-          Current message
-        </TextParagraph>
-
-        <BackgroundCanvas
-          variant="lighterCanvas"
-          className={styles.description}
-        >
-          <TextParagraph variant="whiteText" size="large">
-            {arrayChanges[0].message}
+        <div className={`${styles.sub} ${styles.right}`}>
+          <TextParagraph
+            className={`${styles.defaultSpacing} ${styles.rightSide}`}
+          >
+            Submited
           </TextParagraph>
-        </BackgroundCanvas>
-        <TextParagraph className={styles.smallerSpacing}>
-          Description
-        </TextParagraph>
-
-        <BackgroundCanvas
-          variant="lighterCanvas"
-          className={styles.description}
-        >
-          <TextParagraph variant="whiteText" size="large">
-            {description}
+          <TextBold className={styles.largeSpacing}>
+            {dateFormat(createdAt)}
+          </TextBold>
+        </div>
+      </div>
+      <div className={styles.bottom}>
+        <div>
+          <TextParagraph className={styles.defaultSpacing}>By</TextParagraph>
+          <TextBold className={styles.largeSpacing}>{userName}</TextBold>
+        </div>
+        <div className={`${styles.sub} ${styles.right}`}>
+          <TextParagraph
+            className={`${styles.defaultSpacing} ${styles.rightSide}`}
+          >
+            Location
           </TextParagraph>
-        </BackgroundCanvas>
-        <TextBold>History</TextBold>
-        <TextParagraph className={styles.smallerSpacing}>
-          {arrayChanges.length} updates
+          <TextBold className={styles.largeSpacing}>{location}</TextBold>
+        </div>
+      </div>
+
+      <TextParagraph className={styles.smallerSpacing}>
+        Current message
+      </TextParagraph>
+
+      <BackgroundCanvas variant="lighterCanvas" className={styles.description}>
+        <TextParagraph variant="whiteText" size="large">
+          {arrayChanges[0].message}
         </TextParagraph>
-        <Pagination
-          totalCards={totalCards}
-          cardsPerPage={cardsPerPage}
-          setCurrentPage={setCurrentPage}
-        />
-        {currentCards.map((change) => (
-          <StatusCard
-            key={change.id}
-            issueStatus={change.status}
-            issueMessage={change.message}
-            issueDate={change.createdAt}
-            isHistory={true}
-          />
-        ))}
       </BackgroundCanvas>
-    </div>
+      <TextParagraph className={styles.smallerSpacing}>
+        Description
+      </TextParagraph>
+
+      <BackgroundCanvas variant="lighterCanvas" className={styles.description}>
+        <TextParagraph variant="whiteText" size="large">
+          {description}
+        </TextParagraph>
+      </BackgroundCanvas>
+      <TextBold>History</TextBold>
+      <TextParagraph className={styles.smallerSpacing}>
+        {arrayChanges.length} updates
+      </TextParagraph>
+      <Pagination
+        totalCards={totalCards}
+        cardsPerPage={cardsPerPage}
+        setCurrentPage={setCurrentPage}
+      />
+      {currentCards.map((change) => (
+        <StatusCard
+          key={change.id}
+          issueStatus={change.status}
+          issueMessage={change.message}
+          issueDate={change.createdAt}
+        />
+      ))}
+    </BackgroundCanvas>
   );
 }
