+++ components/Admin/Details/EditBar/EditBar.module.css
@@ -1,0 +1,166 @@
+.topCard {
+  display: flex;
+  justify-content: space-between;
+}
+
+.buttonsPannel {
+  display: none;
+}
+
+.topCardInner {
+  /* display: flex;
+  align-items: center;
+
+  gap: 20px; */
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+/* width: 100%;
+justify-content: space-between; */
+
+.editParts {
+  /* width: 50%; */
+  /* display: flex;
+  justify-content: right;
+  align-items: center; */
+}
+
+.status {
+  color: var(--color--pale-grey);
+}
+
+.pending {
+  align-self: auto;
+}
+
+.hidden {
+  display: none;
+}
+
+/* .editParts {
+  display: none;
+} */
+
+@media screen and (min-width: 767px) {
+  .topCardInner {
+    display: flex;
+    align-items: center;
+    max-width: 28%;
+    gap: 20px;
+    /* width: 100%;
+    justify-content: space-between; */
+  }
+
+  .buttonsPannel {
+    display: flex;
+    gap: 8px;
+  }
+}
+
+/* .editStatusButton {
+  display: flex;
+  align-items: center;
+  height: 100%;
+} */
+
+.deleteButton {
+  background-color: #ec5757 !important;
+}
+
+.solvedButton {
+  background-color: var(--color--intense-purple) !important;
+}
+
+.rightSection {
+  display: flex;
+  align-items: center;
+  height: 100%;
+}
+
+.editButton {
+  padding: 18px 23px 15px 24px;
+  border-radius: 24px;
+  background-color: #252945;
+}
+
+.buttonEdit {
+  border-style: none;
+  background-color: #252945;
+  border-radius: 24px;
+}
+
+@media screen and (max-width: 240px) {
+  .statusCardContainer {
+    display: flex;
+    flex-direction: column;
+  }
+
+  .status {
+    margin-bottom: 1rem;
+  }
+
+  .pending {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+  }
+}
+
+.overlay {
+  background-color: var(--color--bg-canvas);
+  opacity: 0.9;
+  position: fixed;
+  margin-top: -20rem;
+
+  inset: 0;
+  z-index: 30;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.statusCardContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 24px;
+  padding: 24px;
+}
+
+.topCard {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.currentMessage {
+  width: 100%;
+}
+
+.messageCanvas {
+  padding: 24px;
+  text-align: center;
+}
+
+.selectorInner {
+  width: 100px;
+  padding: 14px 0 11px 0;
+  text-align: center;
+  align-self: flex-end;
+  color: var(--color--orange);
+  /* background-color: var(--color--bg-submited); */
+
+  border-radius: 6px;
+}
+
+@media screen and (min-width: 640px) {
+  .statusCardContainer {
+    padding: 24px 48px 27px;
+  }
+}
